backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
PyJWT==2.8.0
bcrypt==4.0.1
python-multipart==0.0.9
redis==5.0.4
httpx==0.27.0
slowapi==0.1.9
